Change to Maven framework to add libraries
