--- src/Main.java
@@ -1,13 +1,0 @@
-package movies.src;
-
-import java.time.LocalDate;
-import movies.src.domain.entities.Customer;
-import movies.src.domain.entities.Movie;
-import movies.src.domain.entities.Rental;
-import movies.src.domain.enums.MovieType;
-
-public class Main {
-    public static void main(String[] args) {
-
-    }
-}
