--- src/domain/entities/Inventory.java
@@ -1,23 +1,0 @@
-package movies.src.domain.entities;
-
-public class Inventory {
-    private final int movieId;
-    private int availableCopies;
-
-    public Inventory(int movieId, int availableCopies) {
-        this.movieId = movieId;
-        this.availableCopies = availableCopies;
-    }
-
-    public int getMovieId() {
-        return movieId;
-    }
-
-    public int getAvailableCopies() {
-        return availableCopies;
-    }
-
-    public void setAvailableCopies(int availableCopies) {
-        this.availableCopies = availableCopies;
-    }
-}
