--- src/domain/exceptions/EntityNotFoundException.java
@@ -1,8 +1,0 @@
-package movies.src.domain.exceptions;
-
-public class EntityNotFoundException extends RuntimeException {
-    public EntityNotFoundException(String message) {
-        super(message);
-    }
-}
-
