--- src/domain/exceptions/InvalidArgumentException.java
@@ -1,7 +1,0 @@
-package movies.src.domain.exceptions;
-
-public class InvalidArgumentException extends IllegalArgumentException {
-    public InvalidArgumentException(String message) {
-        super(message);
-    }
-}
