--- src/domain/exceptions/InvalidRentalOperationException.java
@@ -1,7 +1,0 @@
-package movies.src.domain.exceptions;
-
-public class InvalidRentalOperationException extends RuntimeException {
-    public InvalidRentalOperationException(String message) {
-        super(message);
-    }
-}
