--- src/domain/exceptions/InventoryUnavailableException.java
@@ -1,7 +1,0 @@
-package movies.src.domain.exceptions;
-
-public class InventoryUnavailableException extends RuntimeException {
-    public InventoryUnavailableException(String message) {
-        super(message);
-    }
-}
