--- src/domain/strategies/interfaces/MovieRentalCalculationStrategy.java
@@ -1,6 +1,0 @@
-package movies.src.domain.strategies.interfaces;
-
-public interface MovieRentalCalculationStrategy {
-    double calculateCharge(int daysRented);
-    int calculateFrequentRenterPoints(int daysRented);
-}
