+++ src/main/java/org/rental/Main.java
@@ -1,0 +1,7 @@
+package org.rental;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello, World!");
+    }
+}
