+++ src/main/java/org/rental/domain/entities/Customer.java
@@ -1,0 +1,19 @@
+package org.rental.domain.entities;
+
+public class Customer {
+    private final int id;
+    private final String name;
+
+    public Customer(int id, String name) {
+        this.id = id;
+        this.name = name;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
