+++ src/main/java/org/rental/domain/entities/Movie.java
@@ -1,0 +1,27 @@
+package org.rental.domain.entities;
+
+import org.rental.domain.enums.MovieType;
+
+public class Movie {
+    private final int id;
+    private final String title;
+    private final MovieType movieType;
+
+    public Movie(int id, String title, MovieType movieType) {
+        this.id = id;
+        this.title = title;
+        this.movieType = movieType;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public MovieType getMovieType() {
+        return movieType;
+    }
+}
