+++ src/main/java/org/rental/domain/exceptions/EntityNotFoundException.java
@@ -1,0 +1,8 @@
+package org.rental.domain.exceptions;
+
+public class EntityNotFoundException extends RuntimeException {
+    public EntityNotFoundException(String message) {
+        super(message);
+    }
+}
+
