+++ src/main/java/org/rental/domain/exceptions/InvalidArgumentException.java
@@ -1,0 +1,7 @@
+package org.rental.domain.exceptions;
+
+public class InvalidArgumentException extends IllegalArgumentException {
+    public InvalidArgumentException(String message) {
+        super(message);
+    }
+}
