+++ src/main/java/org/rental/domain/exceptions/InvalidRentalOperationException.java
@@ -1,0 +1,7 @@
+package org.rental.domain.exceptions;
+
+public class InvalidRentalOperationException extends RuntimeException {
+    public InvalidRentalOperationException(String message) {
+        super(message);
+    }
+}
