+++ src/main/java/org/rental/domain/exceptions/InventoryUnavailableException.java
@@ -1,0 +1,7 @@
+package org.rental.domain.exceptions;
+
+public class InventoryUnavailableException extends RuntimeException {
+    public InventoryUnavailableException(String message) {
+        super(message);
+    }
+}
