+++ src/main/java/org/rental/domain/strategies/interfaces/MovieRentalCalculationStrategy.java
@@ -1,0 +1,6 @@
+package org.rental.domain.strategies.interfaces;
+
+public interface MovieRentalCalculationStrategy {
+    double calculateCharge(int daysRented);
+    int calculateFrequentRenterPoints(int daysRented);
+}
